move storage code 1 level up
