+++ src/js/core/methods/AbstractMethod.js
@@ -7,7 +7,7 @@ import DataManager from '../../data/DataManager';
 import * as UI from '../../constants/ui';
 import * as DEVICE from '../../constants/device';
 import * as ERROR from '../../constants/errors';
-import { load as loadStorage, save as saveStorage, PERMISSIONS_KEY } from '../../iframe/storage';
+import { load as loadStorage, save as saveStorage, PERMISSIONS_KEY } from '../../storage';
 
 import { UiMessage, DeviceMessage } from '../../message/builder';
 import type { Deferred, CoreMessage, UiPromiseResponse, FirmwareRange } from '../../types';
